feat: Order cancer types based on their restricted mean survival rate.
Pathways are plotted in different plots. For each plot, the x-axis is stage I - stage IV, and the
y-axis is the mean TPM. Dots are colored red if the gene is significantly differentially expressed
in the stage vs. normals. These significant FC values are then taken together, and for each gene the
max(FC) - min(FC) was calculated to find the "most differentially expressed genes". These genes have
lines colored in red and are labeled with their symbols and log2FC values for each stage.